server/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.12.0
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
Deprecated==1.3.1
exceptiongroup==1.3.1
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.11.0
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==4.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==4.0.1
mdurl==0.1.2
mysqlclient==2.2.7
ordered-set==4.1.0
packaging==24.2
payos==1.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.2.1
requests==2.32.5
rich==13.9.4
SQLAlchemy==2.0.44
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.0.1
zipp==3.23.0
